清理奖励骰及惩罚骰
